+++ CHANGES.md
@@ -4,6 +4,7 @@ videojs-record changelog
 3.5.0 - unreleased
 ------------------
 
+- Support for Picture-in-Picture during playback and recording (#340)
 - New webm-wasm plugin: create webm files using libwebm (compiled with
   WebAssembly) in the browser (#321)
 - Ability to change label of record indicator (#336)